# backend/swagger.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    Full-stack Project & Task Management Platform API.
    
    ## Authentication
    Most endpoints require a JWT access token in the `Authorization: Bearer <token>` header.
    
    ## Roles
    - **ADMIN** — Full access to all resources
    - **PROJECT_MANAGER** — Manage own projects and their tasks
    - **TEAM_MEMBER** — Manage assigned tasks only
  version: 1.0.0
  contact:
    name: TaskFlow Team
servers:
  - url: http://localhost:5000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15min expiry)
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, PROJECT_MANAGER, TEAM_MEMBER]
        isActive:
          type: boolean
        isVerified:
          type: boolean
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                password:
                  type: string
                  minLength: 8
                  description: "Must contain 1 uppercase letter and 1 number"
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Invalid credentials
        '403':
          description: Email not verified
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        '401':
          description: Invalid refresh token
  /auth/profile:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
    put:
      tags: [Auth]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/password:
    put:
      tags: [Auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password updated
        '401':
          description: Current password incorrect
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset link sent (if email exists)
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset successfully
        '401':
          description: Invalid or expired token
  /auth/verify-email/{token}:
    get:
      tags: [Auth]
      summary: Verify email address
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified
        '401':
          description: Invalid or expired token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (invalidate refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logged out
...
  /activity/entity/{type}/{id}:
    get:
      tags: [Activity]
      summary: Get activity for entity
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [TASK, PROJECT, COMMENT]
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entity activity log
  /activity/user/{userId}:
    get:
      tags: [Activity]
      summary: Get user activity history
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User activity log
  /activity/me:
    get:
      tags: [Activity]
      summary: Get current user's activity
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's activity log
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: read
          in: query
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: Paginated notifications with unread count
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
  /notifications/read-all:
    patch:
      tags: [Notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
  /notifications/{id}:
    delete:
      tags: [Notifications]
      summary: Delete notification
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification deleted
  /tasks/{taskId}/attachments:
    get:
      tags: [Attachments]
      summary: List task attachments
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated attachments
    post:
      tags: [Attachments]
      summary: Upload file attachment
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: "Max 10MB. Allowed: images, PDF, Word, Excel, text"
      responses:
        '201':
          description: File uploaded
  /attachments/{id}/download:
    get:
      tags: [Attachments]
      summary: Download attachment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File download
  /attachments/{id}:
    delete:
      tags: [Attachments]
      summary: Delete attachment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attachment deleted